+++ src/storage/index.ts
@@ -3,6 +3,6 @@ export type {
   SessionSnapshot,
   StorageAdapter,
 } from "./storage-adapter.js";
-export { canSaveHeader } from "./storage-adapter.js";
+export { canSaveHeader, StorageConflictError } from "./storage-adapter.js";
 export { FileStorageAdapter } from "./file-storage-adapter.js";
 export { MemoryStorageAdapter } from "./memory-storage-adapter.js";